Deprecated SimpleFactory in favor of MatcherFactory
