+++ tests/ExpandersTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Coduo\PHPMatcher\Tests;
 
-use Coduo\PHPMatcher\Factory\SimpleFactory;
+use Coduo\PHPMatcher\Factory\MatcherFactory;
 use Coduo\PHPMatcher\Matcher;
 use Coduo\PHPMatcher\PHPMatcher;
 use PHPUnit\Framework\TestCase;
@@ -18,7 +18,7 @@ final class ExpandersTest extends TestCase
 
     public function setUp() : void
     {
-        $factory = new SimpleFactory();
+        $factory = new MatcherFactory();
         $this->matcher = $factory->createMatcher();
     }
